backend/requirements.txt
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
pipdeptree==2.26.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
